requirements.txt
boto==2.49.0
boto3==1.9.146
botocore==1.12.146
certifi==2019.3.9
chardet==3.0.4
Click==7.0
docutils==0.14
Flask==1.0.2
gensim==3.7.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
MarkupSafe==1.1.1
nltk==3.4.1
numpy==1.16.3
opencv-python==4.1.0.25
Pillow==6.0.0
pytesseract==0.2.6
python-dateutil==2.8.0
requests==2.21.0
s3transfer==0.2.0
scikit-learn==0.21.0
scipy==1.2.1
six==1.12.0
smart-open==1.8.3
torch==1.0.1.post2
torchvision==0.2.2.post3
urllib3==1.24.3
Werkzeug==0.15.2
